refactor: 프로필 버튼 컴포넌트화 #47
