+++ src/components/layouts/header/teacherHeader.jsx
@@ -1,21 +1,12 @@
-import React, { useState } from 'react';
-import { Avatar, Box, Divider, ListItemIcon, ListItemText, Menu, MenuItem, Typography } from '@mui/material';
-import { AccountBoxRounded, LogoutRounded } from '@mui/icons-material';
+import React from 'react';
+import { Box, Typography } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
 import { useUserAuthStore } from '../../../store';
+import ProfileButton from '../../button/profileButton';
 
 export default function TeacherHeader() {
   const { user } = useUserAuthStore();
   const navigate = useNavigate();
-
-  const [isMouseOver, setMouseOver] = useState(null);
-
-  const handleClick = (e) => {
-    setMouseOver(e.currentTarget);
-  };
-  const handleClose = () => {
-    setMouseOver(null);
-  };
 
   return (
     <Box>
@@ -31,42 +22,7 @@ export default function TeacherHeader() {
         </Typography>
         <Box display="flex" alignItems="center" gap={1}>
           <Typography>강사: {user?.user_name}</Typography>
-          <Avatar onClick={handleClick} />
-          <Menu anchorEl={isMouseOver} open={Boolean(isMouseOver)} onClose={handleClose}>
-            <Box sx={{ paddingX: 2 }}>
-              <Typography variant="subtitle1">
-                {user?.user_name}
-                <br />
-              </Typography>
-              <Typography variant="subtitle2" sx={{ color: 'grey' }}>
-                {user?.email}
-              </Typography>
-            </Box>
-            <Divider />
-            <MenuItem
-              onClick={() => {
-                handleClose();
-                navigate('/teacher/profile');
-              }}
-            >
-              <ListItemIcon>
-                <AccountBoxRounded fontSize="small" />
-              </ListItemIcon>
-              <ListItemText>profile</ListItemText>
-            </MenuItem>
-            <Divider />
-            <MenuItem
-              onClick={() => {
-                handleClose();
-                console.log('Log out');
-              }}
-            >
-              <ListItemIcon>
-                <LogoutRounded fontSize="small" />
-              </ListItemIcon>
-              <ListItemText>Logout</ListItemText>
-            </MenuItem>
-          </Menu>
+          <ProfileButton position="teacher" />
         </Box>
       </Box>
     </Box>
